Absorb backend upload/vertex quirks into VRI; add an upload helper
Three things that leaked backend specifics into example code (found while writing
descriptorindexing) are now handled implicitly, so valid input behaves the same on
every backend:

- Texture upload row-pitch alignment. CmdUploadBufferToTexture took tightly-packed
  data but assumed the app pre-aligned its rows: D3D12 needs a 256-byte row pitch /
  512-byte offset, WebGPU needs bytesPerRow % 256 == 0. A 4x4 texture failed on both
  (worked on VK/GL). Both backends now bounce through a padded buffer when the tight
  pitch isn't aligned (D3D12: CPU memcpy from the mapped staging into a temp upload
  buffer; WebGPU: per-row copyBufferToBuffer into a temp), tracked and freed at the
  next BeginCommandBuffer. Wide/aligned textures keep the direct fast path. D3D12 also
  now selects the array slice from baseLayer (subresource = mip + layer*mipNum).

- Integer vertex attributes. R32/RG32/RGBA32 SINT/UINT were unmapped on GL (always
  glVertexAttribPointer -> the int was read as float garbage) and WebGPU (no Sint32).
  GL now uses glVertexAttribIPointer / glVertexArrayAttribIFormat for integer formats;
  WebGPU maps Sint32/Uint32(x2/x4). So a per-vertex int index just works everywhere.

- ExampleApp upload helper: BeginUpload/UploadBuffer/UploadTexture/EndUpload create the
  staging, record the copies + copy->read barriers, submit/wait, and free the staging -
  replacing the ~40 lines of staging boilerplate each example repeated (and which I once
  got wrong, forgetting a copy -> blank screen).

descriptorindexing now uses all three: an integer texIndex (R32_SINT), tiny 4x4
textures, and the helper. Verified pixel-equal on VK, D3D12, desktop GL, and WebGPU
(wgpu-native, full 16-cell readback), plus WebGL2 + WebGPU on the web. Suite green
(80 / 1865).

+++ tests/shaders/arraytex_wgsl.h
@@ -20,14 +20,14 @@ struct VSOutput_0
 {
     @builtin(position) position_0 : vec4<f32>,
     @location(0) uv_0 : vec2<f32>,
-    @interpolate(flat) @location(1) texIndex_0 : f32,
+    @interpolate(flat) @location(1) texIndex_0 : i32,
 };
 
 struct vertexInput_0
 {
     @location(0) position_1 : vec3<f32>,
     @location(1) uv_1 : vec2<f32>,
-    @location(2) texIndex_1 : f32,
+    @location(2) texIndex_1 : i32,
 };
 
 @vertex
@@ -48,13 +48,13 @@ struct pixelOutput_0
 struct pixelInput_0
 {
     @location(0) uv_2 : vec2<f32>,
-    @interpolate(flat) @location(1) texIndex_2 : f32,
+    @interpolate(flat) @location(1) texIndex_2 : i32,
 };
 
 @fragment
 fn fragmentMain( _S2 : pixelInput_0, @builtin(position) position_2 : vec4<f32>) -> pixelOutput_0
 {
-    var _S3 : vec3<f32> = vec3<f32>(_S2.uv_2, _S2.texIndex_2);
+    var _S3 : vec3<f32> = vec3<f32>(_S2.uv_2, f32(_S2.texIndex_2));
     var _S4 : pixelOutput_0 = pixelOutput_0( (textureSample((uTexArray_0), (uSmp_0), ((_S3)).xy, i32(((_S3)).z))) );
     return _S4;
 }
